requirements.txt
# boto3==1.28.53
# botocore===1.31.53
fastapi==0.73.0
mangum==0.12.3
snowflake-id==0.0.2
requests==2.27.1
starlette==0.17.1
uvicorn==0.17.1
python-socketio==5.11.1
# python-socketio==4.6.0
# python-engineio==3.13.2
pynamodb==6.0.0
aiormq==6.8.0
pika==1.3.2
